pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Technical Research Agent — Plan → Search → Read → Synthesize pipeline with grounded citations."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "langgraph>=0.4.0",
    "anthropic>=0.40.0",
    "pydantic>=2.9",
    "httpx>=0.27",
    "arxiv>=2.1",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "pypdf>=6.10.2",
]

[project.scripts]
research-agent = "research_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/research_agent"]

[tool.hatch.build.targets.wheel.force-include]
"prompts" = "research_agent/prompts"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
